# specs/food_menu_system/openapi.yaml
openapi: 3.0.3
info:
  title: Food Menu System API
  description: REST API for managing food menus in a multi-tenant system.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base URL for all endpoints
paths:
...
  /orders:
    post:
      summary: Create Order
      description: Create a new order for a user.
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
    get:
      summary: List Orders
      description: Retrieve a paginated list of orders for a user or admin.
      tags:
        - Orders
      parameters:
        - $ref: '#/components/parameters/statusQuery'
        - $ref: '#/components/parameters/pageQuery'
        - $ref: '#/components/parameters/limitQuery'
      responses:
        "200":
          description: Orders retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'
...
components:
  schemas:
...
    OrderRequest:
      type: object
      required:
        - menuType
        - menuId
        - items
      properties:
        menuType:
          type: string
          enum:
            - weekly
            - catering
        menuId:
          type: string
          format: uuid
        items:
          type: array
          items:
            type: object
            required:
              - menuItemId
              - quantity
            properties:
              menuItemId:
                type: string
                format: uuid
              quantity:
                type: integer
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        menuType:
          type: string
          enum:
            - weekly
            - catering
        menuId:
          type: string
          format: uuid
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: string
                format: uuid
              quantity:
                type: integer
              price:
                type: number
                format: float
        status:
          type: string
        orderDate:
          type: string
          format: date-time
        deliveryDate:
          type: string
          format: date-time
        totalPrice:
          type: number
          format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        meta:
          type: object
...
  parameters:
...
    statusQuery:
      name: status
      in: query
      required: false
      schema:
        type: string
    pageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
    limitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 10
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT